pyproject.toml
[tool.poetry]
name = "dot"
version = "0.1.0"
description = "Social media application"
authors = ["Henry <hi@henry.moe>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.13"
uvicorn = "^0.34.3"
fastapi = "^0.115.14"
jinja2 = "^3.1.6"
python-dotenv = "^1.1.1"
asyncpg = "^0.30.0"
torch = "^2.7.1"
protobuf = "^6.31.1"
transformers = "^4.53.0"
bcrypt = "^4.3.0"
redis = "^6.2.0"
timeago = "^1.0.16"
bleach = "^6.2.0"
sqlalchemy = {extras = ["asyncpg"], version = "^2.0.41"}
mypy = "^1.16.1"
setuptools = "^80.9.0"
pylint = "^3.3.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
